feat: add database tests for user and db
